requirements.txt
boto3==1.12.26
botocore==1.15.26
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
docutils==0.15.2
idna==2.9
jmespath==0.9.5
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
